rename files using camel case
